Upgrade to new WebGPU EP
* ONNX Runtime improvements (experimental native webgpu; fix iOS) (#1231)

* customize the wasm paths

* update implementation

* allow using 'webgpu' in nodejs binding

* update version of onnxruntime-node

* Upgrade onnxruntime-web to same version as onnxruntime-node

* Update list of supported devices

---------


* customize the wasm paths (#1250)

* customize the wasm paths

* update implementation

* [internal] Add is_decoder option to session retrieval for preferred output location

* Update tests

* Formatting

* Bump ort versions

* Bump onnxruntime-node version

* Bump versions

* Bump ORT versions

* Bump versions

* Only check webgpu fp16 for non-node environments

* Fix

* Assume node supports webgpu

* Update ORT node support comment

* Relax test strictness

* Update conversion script versions

* Downgrade onnxslim

* cleanup

* Update package-lock.json

* Update onnxruntime versions

* Update post-build script

* Use built-in session release function

* Call garbage collection after each tokenizer test

* Do not double-throw error

* Fix race-condition in build process with file removal

* Update versions

* Bump jinja version

* [version] Update to 3.6.3

* Bump jinja version to support new features

* [version] Update to 3.6.3

* Add support for LFM2 models (#1367)

* Use prefix in lfm2 output location (#1369)

* Update package-lock.json

* Run `npm audit fix`

* Add special tokens in text-generation pipeline if tokenizer requires (#1370)

* Add special tokens in text-generation pipeline if tokenizer requires

* Fix logits processors tests

* Update bundles.test.js

* Update comment

* Formatting

* Add support for ModernBERT Decoder (#1371)

* Use from/to buffer instead of string

Actually fixes https://github.com/huggingface/transformers.js/issues/1343

* Add support for Voxtral (#1373)

* Support longform voxtral processing (#1375)

* [version] Update to 3.7.0

* Add support for Arcee (#1377)

* Optimize tensor.slice() (#1381)

* Optimize tensor.slice()

The performance of executing `tensor.slice()` is super poor, especially for
the 'logits' tensor with large dimensions.

```
const logits = outputs.logits.slice(null, -1, null);`
```

This is because currently implementation of the `slice` method manually iterates
through each element and calculate indices which is a big time consuming if
the tensor shape is large.

For cases like `slice(null, -1, null)`, where the slicing operation is
contiguous along certain dimensions, which can be optimized by bulk copy
by using `TypeArray.subarray()` and `TypeArray.set()`.

* nit

* Add a few more tensor slice unit tests

---------


---------

+++ webpack.config.js
@@ -37,24 +37,30 @@ class StripNodePrefixPlugin extends webpack.NormalModuleReplacementPlugin {
  * @see https://webpack.js.org/contribute/writing-a-plugin/
  */
 class PostBuildPlugin {
+  static completed = false;
 
   apply(compiler) {
     compiler.hooks.done.tap('PostBuildPlugin', () => {
-      const dist = path.join(__dirname, 'dist');
-      const ORT_JSEP_FILE = 'ort-wasm-simd-threaded.jsep.mjs';
-      const ORT_BUNDLE_FILE = 'ort.bundle.min.mjs';
-
-      // 1. Remove unnecessary files
-      {
-        const file = path.join(dist, ORT_BUNDLE_FILE);
-        if (fs.existsSync(file)) fs.unlinkSync(file);
+      if (!process.env.WEBPACK_SERVE && !PostBuildPlugin.completed) {
+        // Ensure we only run this once
+        PostBuildPlugin.completed = true;
+        return;
       }
+      const dist = path.join(__dirname, 'dist');
+      const ORT_JSEP_FILE = 'ort-wasm-simd-threaded.asyncify.mjs';
+      const ORT_BUNDLE_FILE = 'ort.webgpu.bundle.min.mjs';
 
-      // 2. Copy unbundled JSEP file
+      // 1. Copy unbundled asyncify file
       {
         const src = path.join(__dirname, 'node_modules/onnxruntime-web/dist', ORT_JSEP_FILE);
         const dest = path.join(dist, ORT_JSEP_FILE);
         fs.copyFileSync(src, dest);
+      }
+
+      // 2. Remove unnecessary files
+      {
+        const file = path.join(dist, ORT_BUNDLE_FILE);
+        if (fs.existsSync(file)) fs.unlinkSync(file);
       }
     });
   }
@@ -184,7 +190,8 @@ const WEB_BUILD = buildConfig({
   ignoreModules: WEB_IGNORE_MODULES,
   externalModules: WEB_EXTERNAL_MODULES,
   plugins: [
-    new StripNodePrefixPlugin()
+    new StripNodePrefixPlugin(),
+    new PostBuildPlugin(),
   ]
 });
